Промежуточный комит ДЗ 19 го спринта
