+++ app/build.gradle.kts
@@ -60,5 +60,8 @@ dependencies {
     implementation ("androidx.fragment:fragment-ktx:1.5.5")
     implementation ("androidx.viewpager2:viewpager2:1.0.0")
     implementation ("com.google.android.material:material:1.8.0")
+    implementation ("androidx.navigation:navigation-fragment-ktx:2.5.3")
+    implementation ("androidx.navigation:navigation-ui-ktx:2.5.3")
+    implementation ("androidx.fragment:fragment-ktx:1.5.6")
 
 }
