+++ app/src/main/java/com/practicum/playlistmaker/data/storage/IntentWork.kt
@@ -8,7 +8,6 @@ import com.practicum.playlistmaker.data.App
 import com.practicum.playlistmaker.domain.model.Track
 import com.practicum.playlistmaker.presentation.UI.Media.MediaActivity
 import com.practicum.playlistmaker.presentation.UI.PlayerActivity
-import com.practicum.playlistmaker.presentation.UI.SearchActivityNew
 import com.practicum.playlistmaker.presentation.UI.SettingsActivity
 
 class IntentWork(private val gson: Gson, private val myApp: App) {
@@ -39,9 +38,9 @@ class IntentWork(private val gson: Gson, private val myApp: App) {
          myApp.startActivity(intent)
     }
      fun openSearch() {
-         val intSearch = Intent(myApp, SearchActivityNew::class.java)
-         intSearch.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-         myApp.startActivity(intSearch)
+//         val intSearch = Intent(myApp, SearchActivityNew::class.java)
+//         intSearch.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+//         myApp.startActivity(intSearch)
     }
 
      fun openMedia() {
