+++ app/src/main/java/com/practicum/playlistmaker/presentation/UI/Media/ParrentMediaFragment.kt
@@ -24,7 +24,7 @@ class ParrentMediaFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-
+        super.onCreateView(inflater, container, savedInstanceState)
         binding = ParrentMediaFragmentBinding.inflate(inflater, container, false)
         return binding!!.root
     }
