+++ app/src/main/java/com/practicum/playlistmaker/presentation/UI/ParentActivity.kt
@@ -1,0 +1,31 @@
+package com.practicum.playlistmaker.presentation.UI
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.setupWithNavController
+import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.practicum.playlistmaker.R
+import com.practicum.playlistmaker.databinding.ActivityParentBinding
+import com.practicum.playlistmaker.presentation.ViewModels.ParentViewModel
+import org.koin.androidx.viewmodel.ext.android.viewModel
+
+class ParentActivity : AppCompatActivity() {
+
+    private val viewModel: ParentViewModel by viewModel()
+    private lateinit var binding: ActivityParentBinding
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_parent)
+
+        binding = ActivityParentBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        val navHostFragment = supportFragmentManager.findFragmentById(R.id.rootFragmentContainerView) as NavHostFragment
+        val navController = navHostFragment.navController
+
+        val bottomNavigationView = findViewById<BottomNavigationView>(R.id.bottom_nav_view)
+        bottomNavigationView.setupWithNavController(navController)
+
+    }
+}
