+++ app/src/main/java/com/practicum/playlistmaker/presentation/ViewModels/ParentViewModel.kt
@@ -1,0 +1,10 @@
+package com.practicum.playlistmaker.presentation.ViewModels
+
+import androidx.lifecycle.ViewModel
+
+
+class ParentViewModel(): ViewModel() {
+
+
+
+}
