+++ app/src/main/res/layout/settings_fragment.xml
@@ -1,0 +1,103 @@
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    style="@style/mySetBackColor"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+
+    <LinearLayout
+        style="@style/LinearLayoutStyle"
+        android:layout_marginStart="@dimen/dp_20"
+        android:layout_marginTop="@dimen/dp_20"
+        tools:ignore="UseCompoundDrawables">
+
+        <ImageView
+            android:id="@+id/iv_back"
+            style="@style/ImageViewStyle2"
+            android:layout_gravity="center"
+             />
+
+        <TextView
+            style="@style/mySetTextColor"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:layout_marginStart="@dimen/dp_20"
+            android:fontFamily="@font/ys_display_medium"
+            android:text="@string/button_settings"
+            android:textFontWeight="500"
+            android:textSize="@dimen/sp_22" />
+
+    </LinearLayout>
+
+    <LinearLayout
+        style="@style/LinearLayoutStyle"
+        android:layout_marginTop="@dimen/dp_60"
+        android:layout_marginEnd="@dimen/dp_10">
+
+        <TextView
+            style="@style/TextViewStyle"
+            android:layout_gravity="center"
+            android:text="@string/text_nt" />
+
+        <Switch
+            android:id="@+id/sw_theme"
+            android:layout_width="@dimen/dp_44"
+            android:layout_height="@dimen/dp_44"
+            android:layout_gravity="end"
+            tools:ignore="UseSwitchCompatOrMaterialXml" />
+
+    </LinearLayout>
+
+    <LinearLayout
+        style="@style/LinearLayoutStyle"
+        android:layout_marginTop="35dp"
+        android:layout_marginEnd="@dimen/dp_20"
+        tools:ignore="UseCompoundDrawables">
+
+        <TextView
+            style="@style/TextViewStyle"
+            android:text="@string/text_share" />
+
+        <ImageView
+            android:id="@+id/iv_share"
+            style="@style/ImageViewStyle"
+            android:src="@drawable/share" />
+
+    </LinearLayout>
+
+    <LinearLayout
+        style="@style/LinearLayoutStyle"
+        android:layout_marginTop="@dimen/dp_45"
+        android:layout_marginEnd="@dimen/dp_20"
+        tools:ignore="UseCompoundDrawables">
+
+        <TextView
+            style="@style/TextViewStyle"
+            android:text="@string/text_support" />
+
+        <ImageView
+            android:id="@+id/iv_support"
+            style="@style/ImageViewStyle"
+            android:src="@drawable/support" />
+
+    </LinearLayout>
+
+    <LinearLayout
+        style="@style/LinearLayoutStyle"
+        android:layout_marginTop="@dimen/dp_45"
+        android:layout_marginEnd="@dimen/dp_20"
+        tools:ignore="UseCompoundDrawables">
+
+        <TextView
+            style="@style/TextViewStyle"
+            android:text="@string/text_userallow" />
+
+        <ImageView
+            android:id="@+id/iv_allow"
+            style="@style/ImageViewStyle"
+            android:src="@drawable/allow" />
+    </LinearLayout>
+
+</LinearLayout>
