+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/searchFragment"
+        android:title="TODO" />
+
+    <item
+        android:id="@+id/parrentMediaFragment"
+        android:title="TODO" />
+
+    <item
+        android:id="@+id/settingsFragment"
+        android:title="TODO" />
+
+</menu>
