+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/main_nav_graph"
+    app:startDestination="@id/searchFragment">
+
+    <fragment
+        android:id="@+id/searchFragment"
+        android:name="com.practicum.playlistmaker.presentation.UI.SearchFragment"
+        android:label="SearchFragment" >
+        <action
+            android:id="@+id/action_searchFragment_to_playerActivity"
+            app:destination="@id/playerActivity" />
+    </fragment>
+    <activity
+        android:id="@+id/playerActivity"
+        android:name="com.practicum.playlistmaker.presentation.UI.PlayerActivity"
+        android:label="PlayerActivity" />
+    <fragment
+        android:id="@+id/parrentMediaFragment"
+        android:name="com.practicum.playlistmaker.presentation.UI.Media.ParrentMediaFragment"
+        android:label="ParrentMediaFragment" />
+    <fragment
+        android:id="@+id/settingsFragment"
+        android:name="com.practicum.playlistmaker.presentation.UI.SettingsFragment"
+        android:label="SettingsFragment" />
+</navigation>
